web/requirements.txt
streamlit==1.54.0
streamlit-authenticator==0.4.2
pandas==2.3.3
pyarrow==23.0.1
streamlit-folium==0.26.1
cloud-sql-python-connector[pg8000]==1.20.0
sqlalchemy==2.0.46
google-cloud-storage==3.9.0